More Cleanup:
added Product ID props to componenets
Removed getStartingPid.jsx
Decided to use jsx files (trading the cleaner looking filepath includes
for fewer errors later.  We can always rename them back to .js if we
don't need them as .jsx after everything is done!)
