+++ client/src/index.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import ReactDOM from 'react-dom';
 import axios from 'axios';
-import getStartingPid from './getStartingPid.jsx'
+//import getStartingPid from './getStartingPid.jsx'
 import Overview from './overview/index.jsx'
 import Questions from './questions/index.jsx'
 import Reviews from './reviews/index.jsx'
@@ -13,10 +13,11 @@ class App extends React.Component {
       product_id: null
     };
     this.singleItemRequest = this.singleItemRequest.bind(this);
+    this.getStartingPid = this.getStartingPid.bind(this);
   }
 
   componentDidMount() {
-    getStartingPid()
+    this.getStartingPid()
     .then((id) => {
       this.setState({
         product_id: id
@@ -47,12 +48,29 @@ class App extends React.Component {
       })
   }
 
+ getStartingPid() {
+    var config = {
+      method: 'get',
+      url: 'http://localhost:10038/starting_product_id'
+    };
+
+    return axios(config)
+      .then((itemsResponse) => {
+        return new Promise( (resolve, reject) => {
+          resolve( itemsResponse.data.product_id );
+        });
+      })
+      .catch((error) => {
+        console.log(error);
+      })
+  }
+
   render () {
     return (
       <div>
-        <Overview />
-        <Questions />
-        <Reviews />
+        <Overview product_id={this.state.product_id} />
+        <Questions product_id={this.state.product_id} />
+        <Reviews product_id={this.state.product_id} />
       </div>
     )
   }
